cleanup: merge crawler scripts
