+++ tests/test_regex.py
@@ -1,6 +1,6 @@
 import os, sys; sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 import re
-from tuchtrecht_crawler import TuchtrechtCrawler
+from main import TuchtrechtCrawler
 
 def test_xml_pattern():
     pattern = TuchtrechtCrawler.XML_PATTERN
